Prevent error inserting large MMSI numbers into the dirty table.
Error caused by large MMSIs (> maxint) being inserted into dirty table. In Python all int = long so these values are not truncated in the parsing stage, but will be rejected by the MMSI validation. We allow MMSIs in the dirty table to be bigint so such large values can still be inserted.

+++ pyrate/repositories/aisdb.py
@@ -53,7 +53,7 @@ class AISdb(sql.PgsqlRepository):
 
     dirty_db_spec = {
         'cols': [
-            ('MMSI', 'integer'),
+            ('MMSI', 'bigint'),
             ('Time', 'timestamp without time zone'),
             ('Message_ID', 'integer'),
             ('Navigational_status', 'integer'),
